requirements.txt
opencv-python-headless
tensorflow
numpy
pillow
matplotlib
streamlit
